fix signed int overflow in nonce rand

+++ ccminer.cpp
@@ -1957,7 +1957,7 @@ static void *miner_thread(void *userdata)
 
 		if (opt_algo == ALGO_EQUIHASH) {
 		    //nonceptr[1] = (rand()*4);
-			nonceptr[2] = rand() << 24 | rand() << 8 | thr_id;
+			nonceptr[2] = (uint32_t)rand() << 24 | (uint32_t)rand() << 8 | thr_id;
 			//applog_hex(&work.data[27], 32);
 		}
